feat(admin-ai-settings): implement shared validation for providerOptions JSON and refactor validators
