+++ apps/app/src/features/mastra/server/routes/admin-ai-settings/put-ai-settings.spec.ts
@@ -38,6 +38,8 @@ vi.mock(
   }),
 );
 
+import type { Request } from 'express';
+import { validationResult } from 'express-validator';
 import { mock } from 'vitest-mock-extended';
 
 import { SupportedAction } from '~/interfaces/activity';
@@ -46,7 +48,10 @@ import type Crowi from '~/server/crowi';
 import type { ApiV3Response } from '~/server/routes/apiv3/interfaces/apiv3-response';
 
 import type { AiSettingsUpdateRequest } from '../../../interfaces/ai-settings';
-import { putAiSettingsFactory } from './put-ai-settings';
+import {
+  putAiSettingsFactory,
+  updateAiSettingsValidators,
+} from './put-ai-settings';
 
 const ACTIVITY_ID = 'activity-id-1';
 
@@ -226,5 +231,148 @@ describe('putAiSettings (Req 2.3, 2.4, 4.3, 4.4, 5.3, 7.1)', () => {
       // The audit event must not fire on failure.
       expect(emit).not.toHaveBeenCalled();
     });
+  });
+});
+
+// --- updateAiSettingsValidators (Req 6.1, 6.2) -----------------------------
+//
+// The validator chain enforces the FORMAL request shape before the handler runs.
+// We assert its observable contract (accept / reject of a field) by driving the
+// real express-validator engine over a fake request and inspecting validationResult,
+// rather than the chain's internal structure.
+
+// Build a minimal Express-like request the express-validator engine accepts.
+// Only `body` carries real data; the other locations are present so the engine
+// can traverse them without throwing.
+const buildRequest = (body: Record<string, unknown>): Request =>
+  ({
+    body,
+    cookies: {},
+    headers: {},
+    params: {},
+    query: {},
+  }) as unknown as Request;
+
+// Run the full chain against a request body and report whether the engine
+// accumulated errors, plus which fields failed.
+const runValidators = async (
+  body: Record<string, unknown>,
+): Promise<{ hasErrors: boolean; failedFields: string[] }> => {
+  const req = buildRequest(body);
+  await Promise.all(updateAiSettingsValidators.map((chain) => chain.run(req)));
+  const result = validationResult(req);
+  return {
+    hasErrors: !result.isEmpty(),
+    failedFields: result.array().map((e) => e.param),
+  };
+};
+
+describe('updateAiSettingsValidators (Req 6.1, 6.2)', () => {
+  describe('provider', () => {
+    it.each([
+      'openai',
+      'anthropic',
+      'google',
+      'azure-openai',
+    ])('accepts the supported provider "%s"', async (provider) => {
+      const { hasErrors } = await runValidators({ provider });
+      expect(hasErrors).toBe(false);
+    });
+
+    it('rejects an unsupported provider value', async () => {
+      const { hasErrors, failedFields } = await runValidators({
+        provider: 'bedrock',
+      });
+      expect(hasErrors).toBe(true);
+      expect(failedFields).toContain('provider');
+    });
+
+    it('accepts a request that omits provider (optional / partial update)', async () => {
+      const { hasErrors } = await runValidators({});
+      expect(hasErrors).toBe(false);
+    });
+  });
+
+  // providerOptions now uses the shared FE/BE predicate (isValidProviderOptionsJson,
+  // JSON.parse based), so client and server accept/reject exactly the same input.
+  describe('providerOptions (shared JSON.parse predicate)', () => {
+    it('accepts a parsable JSON object string', async () => {
+      const { hasErrors } = await runValidators({
+        providerOptions: '{"openai":{"temperature":0.7}}',
+      });
+      expect(hasErrors).toBe(false);
+    });
+
+    // JSON.parse accepts objects, arrays, the literal primitives true/false/null,
+    // AND bare numbers / quoted strings, so the shared predicate accepts them all.
+    // This is the parity point with the client form.
+    it.each([
+      '[1,2,3]',
+      'true',
+      'false',
+      'null',
+      '42',
+      '"x"',
+    ])('accepts the parsable JSON value "%s"', async (providerOptions) => {
+      const { hasErrors } = await runValidators({ providerOptions });
+      expect(hasErrors).toBe(false);
+    });
+
+    it('rejects a malformed JSON string', async () => {
+      const { hasErrors, failedFields } = await runValidators({
+        providerOptions: '{ invalid',
+      });
+      expect(hasErrors).toBe(true);
+      expect(failedFields).toContain('providerOptions');
+    });
+
+    it('accepts an empty string (cleared option, valid "no options")', async () => {
+      const { hasErrors } = await runValidators({ providerOptions: '' });
+      expect(hasErrors).toBe(false);
+    });
+
+    it('accepts a request that omits providerOptions', async () => {
+      const { hasErrors } = await runValidators({});
+      expect(hasErrors).toBe(false);
+    });
+  });
+
+  describe('boolean fields', () => {
+    it.each([
+      'aiEnabled',
+      'azureOpenaiUseEntraId',
+    ])('accepts a boolean value for "%s"', async (field) => {
+      const accepted = await runValidators({ [field]: true });
+      expect(accepted.hasErrors).toBe(false);
+
+      const acceptedFalse = await runValidators({ [field]: false });
+      expect(acceptedFalse.hasErrors).toBe(false);
+    });
+
+    it.each([
+      'aiEnabled',
+      'azureOpenaiUseEntraId',
+    ])('rejects a non-boolean value for "%s"', async (field) => {
+      const { hasErrors, failedFields } = await runValidators({
+        [field]: 'yes',
+      });
+      expect(hasErrors).toBe(true);
+      expect(failedFields).toContain(field);
+    });
+  });
+
+  it('accepts a fully populated valid request', async () => {
+    const { hasErrors } = await runValidators({
+      aiEnabled: true,
+      provider: 'azure-openai',
+      apiKey: 'secret-key',
+      model: 'gpt-4o',
+      providerOptions: '{"temperature":0.2}',
+      azureOpenaiResourceName: 'my-resource',
+      azureOpenaiBaseUrl: 'https://example.openai.azure.com',
+      azureOpenaiApiVersion: '2024-02-01',
+      azureOpenaiUseEntraId: false,
+    });
+    expect(hasErrors).toBe(false);
   });
 });
